chore: utils>date.ts 파일 삭제, supabase 타입 분리, 세션 없을 시 좋아요 눌렀을 때 알럿, isgoingout props 수정
