+++ src/components/chat/market/ChatMessages.styled.ts
@@ -36,20 +36,20 @@ export const DateLabel = styled.div`
 export const MessageWrapper = styled.div<MessageProps>`
   display: flex;
   flex-direction: column;
-  align-items: ${(props) => (props.isOutgoing ? 'flex-end' : 'flex-start')};
+  align-items: ${(props) => (props.$isoutgoing ? 'flex-end' : 'flex-start')};
 `;
 
 export const MessageComponent = styled.div<MessageComponentProps>`
   display: flex;
   flex-direction: column;
   word-wrap: break-word;
-  background-color: ${(props) => (props.isoutgoing ? '#FCA311' : '#14213D')}; /* 배경 색상을 변경합니다. */
-  color: ${(props) => (props.isoutgoing ? '#fff' : '#fff')}; /* 텍스트 색상을 변경합니다. */
-  align-self: ${(props) => (props.isoutgoing ? 'flex-end' : 'flex-start')};
+  background-color: ${(props) => (props.$isoutgoing ? '#FCA311' : '#14213D')}; /* 배경 색상을 변경합니다. */
+  color: ${(props) => (props.$isoutgoing ? '#fff' : '#fff')}; /* 텍스트 색상을 변경합니다. */
+  align-self: ${(props) => (props.$isoutgoing ? 'flex-end' : 'flex-start')};
   /* 추가: 메시지 버블 안에 텍스트가 중앙에 오도록 만듭니다. */
-  align-items: ${(props) => (props.isoutgoing ? 'flex-end' : 'flex-start')};
+  align-items: ${(props) => (props.$isoutgoing ? 'flex-end' : 'flex-start')};
   font-weight: bold;
-  text-align: ${(props) => (props.isoutgoing ? 'right' : 'left')};
+  text-align: ${(props) => (props.$isoutgoing ? 'right' : 'left')};
   text-align: right;
   padding: 0.6rem;
   margin: 0.3rem;
