+++ src/components/chat/market/ChatMessages.type.ts
@@ -1,8 +1,9 @@
 // MessageComponentProps 인터페이스 정의
 export type MessageComponentProps = {
-  isoutgoing?: string; // 또는 `boolean`이나 다른 타입으로 정의할 수 있습니다, 여기서는 오류 메시지에 맞춰 `string`으로 가정합니다.
+  $isoutgoing?: string; // 또는 `boolean`이나 다른 타입으로 정의할 수 있습니다, 여기서는 오류 메시지에 맞춰 `string`으로 가정합니다.
 };
 
 export type MessageProps = {
-  isOutgoing: boolean;
+  // isoutgoing: boolean;
+  $isoutgoing: string;
 };
