+++ src/components/profile/UserProfile.tsx
@@ -2,10 +2,10 @@ import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { getUserDataHandler } from '../../api/Supabase.api';
+import { Tables } from '../../api/Supabase.type';
 import profileImage from '../../assets/profile/defaultprofileimage.webp';
 import { useUserDataQuery } from '../../hooks/users/useUserData';
 import { userSession } from '../../state/atom/userSessionAtom';
-import { Tables } from '../../types/types';
 import Follow from '../common/follow/Follow';
 import * as St from './UserProfile.styled';
 import Tab from './tab/Tab';
@@ -17,7 +17,6 @@ const UserProfile = () => {
   const session = useRecoilValue(userSession);
 
   const data = useUserDataQuery(params as string);
-  // console.log(data);
 
   const getUserData = async () => {
     const result = await getUserDataHandler(params as string);
