--- src/utils/readme
@@ -1,1 +1,0 @@
-이 폴더 안에는 날짜 형식이 들어갈 예정입니다.
